Merging latest of 1.3-rel into easysetup branch that is lying unused
from 1 year.

We plan to use this easysetup branch as a dev branch for upcoming
patches and merge back into 1.3-rel when its stable to be
released.

Conflicts:
	service/easy-setup/sampleapp/mediator/android/EasySetup/app/src/main/java/org/iotivity/service/easysetup/MainActivity.java

+++ resource/csdk/connectivity/inc/caedradapter_singlethread.h
@@ -129,7 +129,7 @@ int32_t CASendEDRMulticastData(const void *data, uint32_t dataLength);
  * @retval #CA_STATUS_FAILED Operation failed
  * @remarks info is allocated in this API and should be freed by the caller.
  */
-CAResult_t CAGetEDRInterfaceInformation(CAEndpoint_t **info, uint32_t *size);
+CAResult_t CAGetEDRInterfaceInformation(CAEndpoint_t **info, size_t *size);
 
 /**
  * @brief  Read Synchronous API callback.
